chore(content-type): drop the bulk-repair endpoint; rely on on-view self-heal
The manual POST /v1/maintenance/repair-content-types (+ repair lib) is redundant
now that every /raw view self-heals a mislabeled version. Remove it; keep the
publish-time sniff, the render-time passthrough, the on-view reclassify, and
MetaStore.reclassifyVersion (used by the self-heal). The publish-sniff test moves
to content-type-sniff.test.ts; render + self-heal stay covered by raw-self-heal.test.ts.
